Replaced junit parser with refactored
